+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export * from './v2/AreaChart';
+export * from './v2/ApyChart';
 export * from './v2/AuthModal';
 export * from './v2/Button';
 export * from './v2/BscLink';
