+++ src/components/v2/ApyChart/index.tsx
@@ -1,40 +1,47 @@
 /** @jsxImportSource @emotion/react */
 import React from 'react';
-import BigNumber from 'bignumber.js';
-import {
-  AreaChart,
-  Tooltip,
-  Area,
-  CartesianGrid,
-  XAxis,
-  YAxis,
-  ResponsiveContainer,
-} from 'recharts';
+import { AreaChart, Tooltip, Area, CartesianGrid, XAxis, YAxis } from 'recharts';
 import { useUID } from 'react-uid';
-import Typography from '@mui/material/Typography';
-
-import { useTranslation } from 'translation';
-import { formatToReadablePercentage, formatCentsToReadableValue } from 'utilities/common';
-import formatToReadableDate from './formatToReadableDate';
-import { useStyles } from './styles';
-
-export interface IItem {
-  apyPercentage: number;
-  timestamp: Date;
-  balanceCents: BigNumber;
-}
 
 export interface IApyChartProps {
-  data: IItem[];
-  type: 'supply' | 'borrow';
   className?: string;
 }
 
-export const ApyChart: React.FC<IApyChartProps> = ({ className, data, type }) => {
-  const styles = useStyles();
-  const chartColor = type === 'supply' ? styles.supplyChartColor : styles.borrowChartColor;
-  const { Trans } = useTranslation();
+const data = [
+  {
+    name: 'Page A',
+    uv: 40,
+  },
+  {
+    name: 'Page B',
+    uv: 30,
+  },
+  {
+    name: 'Page C',
+    uv: 20,
+  },
+  {
+    name: 'Page D',
+    uv: 27,
+  },
+  {
+    name: 'Page E',
+    uv: 18,
+  },
+  {
+    name: 'Page F',
+    uv: 23,
+  },
+  {
+    name: 'Page G',
+    uv: 34,
+  },
+];
 
+// TODO: pass data through props
+// TODO: get colors from theme
+
+export const ApyChart: React.FC<IApyChartProps> = ({ className }) => {
   // Generate base ID that won't change between renders but will be incremented
   // automatically every time it is used (so multiple charts can be rendered
   // using unique ids)
@@ -42,108 +49,39 @@ export const ApyChart: React.FC<IApyChartProps> = ({ className, data, type }) =>
   const gradientId = `gradient-${baseId}`;
 
   return (
-    <div css={styles.container} className={className}>
-      <ResponsiveContainer>
-        <AreaChart margin={styles.areaChartMargin} data={data}>
-          {/* Gradient used as filler */}
-          <defs>
-            <linearGradient id={gradientId} x1="0" y1="0" x2="0" y2="1">
-              <stop offset="0%" stopColor={chartColor} stopOpacity={0.3} />
-              <stop offset="100%" stopColor={chartColor} stopOpacity={0} />
-            </linearGradient>
-          </defs>
+    <div className={className}>
+      <AreaChart
+        width={700}
+        height={350}
+        data={data}
+        margin={{
+          top: 10,
+          right: 30,
+          left: 0,
+          bottom: 0,
+        }}
+      >
+        {/* Gradient used as filler */}
+        <defs>
+          <linearGradient id={gradientId} x1="0" y1="0" x2="0" y2="1">
+            <stop offset="0%" stopColor="#8884d8" stopOpacity={0.35} />
+            <stop offset="100%" stopColor="#8884d8" stopOpacity={0} />
+          </linearGradient>
+        </defs>
 
-          <CartesianGrid vertical={false} stroke={styles.gridLineColor} />
-          <XAxis
-            dataKey="timestamp"
-            axisLine={false}
-            tickLine={false}
-            tickFormatter={formatToReadableDate}
-            stroke={styles.accessoryColor}
-            tickMargin={styles.tickMargin}
-            style={styles.axis}
-          />
-          <YAxis
-            axisLine={false}
-            tickLine={false}
-            tickFormatter={formatToReadablePercentage}
-            tickMargin={styles.tickMargin}
-            stroke={styles.accessoryColor}
-            style={styles.axis}
-            domain={[0, 'dataMax + 10']}
-          />
-          <Tooltip
-            isAnimationActive={false}
-            cursor={styles.cursor}
-            content={({ payload }) =>
-              payload && payload[0] ? (
-                <div css={styles.tooltipContainer}>
-                  <div css={styles.tooltipItem}>
-                    <Trans
-                      i18nKey={
-                        // Translation keys: do not remove this comment
-                        // t('apyChart.tooltipItems.supplyApy')
-                        // t('apyChart.tooltipItems.borrowApy')
-                        type === 'supply'
-                          ? 'apyChart.tooltipItems.supplyApy'
-                          : 'apyChart.tooltipItems.borrowApy'
-                      }
-                      components={{
-                        Label: <Typography css={styles.tooltipItemLabel} variant="tiny" />,
-                        Value: <Typography css={styles.tooltipItemValue} variant="small1" />,
-                      }}
-                      values={{
-                        percentage: formatToReadablePercentage(
-                          (payload[0].payload as IItem).apyPercentage,
-                        ),
-                      }}
-                    />
-                  </div>
-
-                  <div css={styles.tooltipItem}>
-                    <Trans
-                      i18nKey={
-                        // Translation keys: do not remove this comment
-                        // t('apyChart.tooltipItems.totalSupply')
-                        // t('apyChart.tooltipItems.totalBorrow')
-                        type === 'supply'
-                          ? 'apyChart.tooltipItems.totalSupply'
-                          : 'apyChart.tooltipItems.totalBorrow'
-                      }
-                      components={{
-                        Label: <Typography css={styles.tooltipItemLabel} variant="tiny" />,
-                        Value: <Typography css={styles.tooltipItemValue} variant="small1" />,
-                      }}
-                      values={{
-                        balance: formatCentsToReadableValue({
-                          value: (payload[0].payload as IItem).balanceCents,
-                          shorthand: true,
-                        }),
-                      }}
-                    />
-                  </div>
-                </div>
-              ) : null
-            }
-          />
-          <Area
-            isAnimationActive={false}
-            dataKey="apyPercentage"
-            stroke={chartColor}
-            strokeWidth={styles.areaStrokeWidth}
-            fillOpacity={1}
-            fill={`url(#${gradientId})`}
-            activeDot={styles.areaActiveDot}
-          />
-        </AreaChart>
-      </ResponsiveContainer>
+        <CartesianGrid vertical={false} stroke="#383944" />
+        <XAxis dataKey="name" axisLine={false} tickLine={false} stroke="#9597A1" />
+        {/* TODO: set domain based on data (with maximum starting at 100) */}
+        <YAxis axisLine={false} tickLine={false} stroke="#9597A1" domain={[0, 50]} />
+        <Tooltip isAnimationActive={false} cursor={{ strokeDasharray: '4 4', stroke: '#9597A1' }} />
+        <Area
+          dataKey="uv"
+          stroke="#8884d8"
+          strokeWidth="2px"
+          fillOpacity={1}
+          fill={`url(#${gradientId})`}
+        />
+      </AreaChart>
     </div>
   );
 };
-
-export const SupplyApyChart: React.FC<Omit<IApyChartProps, 'type'>> = props => (
-  <ApyChart type="supply" {...props} />
-);
-export const BorrowApyChart: React.FC<Omit<IApyChartProps, 'type'>> = props => (
-  <ApyChart type="borrow" {...props} />
-);
